Refactoring of faderwidgets
